refactor: 예약 엔티티 수정
- 예약 엔티티 수정 사항
  - 예약 시간
    - 날짜와 시간을 같이 반환하여 사용하던 방식을 분리하여 사용하는
      방식으로 변경
  - 상태
    - 예약 상태, 도착 상태와 승인 상태로 구성되었던 상태 객체의 정보를
      예약 상태와 승인 상태를 예약 상태로 통합
  - 예약 인원 추가
    - 방문을 희망하는 인원 수 정보 추가
